add layout for transference1 refactor steps add BoxHeader rename file Combobox to ComboBox
