+++ src/components/atoms/ComboBox.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const ComboBox = ()=>(
+  <select className="form-control">
+    <option>Selecciona una cuenta</option>
+    <option>Cuenta 1</option>
+    <option>Cuenta 2</option>
+    <option>Cuenta 3</option>
+    </select>
+)
+export default ComboBox;
+
