--- src/components/atoms/Combobox.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-
-const ComboBox = ()=>(
-<select className="form-control">
-  <option>Selecciona una cuenta</option>
-  <option>Cuenta 1</option>
-  <option>Cuenta 2</option>
-  <option>Cuenta 3</option>
-</select>
-)
-export default ComboBox;
-
