+++ src/routes.js
@@ -2,14 +2,13 @@ import React from 'react';
 import { Route , IndexRoute } from 'react-router';
 
 import LoginUser from './components/templates/login/LoginUser';
+import Transference1 from './components/templates/transferences/Transference1.js';
 
 //components
 import App from './App';
-import Test1 from './components/test1';
-import Test2 from './components/test2';
 
-export const getRoutes = () => 
+export const getRoutes = () =>
   <Route path="/" component={App}>
     <IndexRoute component={LoginUser}/>
-    <Route path="test2" component={Test2}/>
+    <Route path="transference1" component={Transference1}/>
   </Route>
